refactor(linter/plugins): split tokens initialization from tokens methods (#19981)
Pure refactor. `tokens.ts` was getting too large. Split it into 2 files:

* `tokens.ts`: Tokens lifecycle methods - initialization, deserialization, and reset.
* `tokens_methods.ts`: Public methods on `SourceCode` e.g. `sourceCode.getFirstToken`.
