+++ apps/oxlint/src-js/plugins.ts
@@ -24,10 +24,6 @@ export type {
 export type { Settings } from "./plugins/settings.ts";
 export type { SourceCode } from "./plugins/source_code.ts";
 export type {
-  CountOptions,
-  FilterFn,
-  RangeOptions,
-  SkipOptions,
   Token,
   BooleanToken,
   IdentifierToken,
@@ -42,6 +38,12 @@ export type {
   StringToken,
   TemplateToken,
 } from "./plugins/tokens.ts";
+export type {
+  CountOptions,
+  FilterFn,
+  RangeOptions,
+  SkipOptions,
+} from "./plugins/tokens_methods.ts";
 export type {
   RuleMeta,
   RuleDocs,
