+++ apps/oxlint/src-js/plugins/source_code.ts
@@ -19,7 +19,7 @@ import { resetScopeManager, SCOPE_MANAGER } from "./scope.ts";
 import * as scopeMethods from "./scope.ts";
 import { resetTokens } from "./tokens.ts";
 import { tokens, tokensAndComments, initTokens, initTokensAndComments } from "./tokens.ts";
-import * as tokenMethods from "./tokens.ts";
+import * as tokenMethods from "./tokens_methods.ts";
 import { debugAssertIsNonNull } from "../utils/asserts.ts";
 
 import type { Program } from "../generated/types.d.ts";
